second commit

+++ Restaurant_API/RestuarantApp_API/AdminController/BranchesController.cs
@@ -9,6 +9,9 @@ using System.Web;
 using System.Web.Mvc;
 using RestuarantApp_API.Models;
 using RestuarantApp_API.Context;
+using RestuarantApp_API.ImageViewModel;
+using RestuarantApp_API.Constants;
+using System.IO;
 
 namespace RestuarantApp_API.AdminController
 {
@@ -19,6 +22,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Branches/
         public async Task<ActionResult> Index()
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             var branches = db.Branches.Include(b => b.Restaurant);
             return View(await branches.ToListAsync());
         }
@@ -26,6 +33,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Branches/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -41,6 +52,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Branches/Create
         public ActionResult Create()
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             ViewBag.RestaurantId = new SelectList(db.Restaurants, "Id", "Name");
             return View();
         }
@@ -50,22 +65,62 @@ namespace RestuarantApp_API.AdminController
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include="Id,Name,Address,ImageUrl,RestaurantId")] BranchInfo branchinfo)
+        public async Task<ActionResult> Create(ViewBranches viewBranches)
         {
-            if (ModelState.IsValid)
+            if (Session["UserName"] == null)
             {
-                db.Branches.Add(branchinfo);
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                return RedirectToAction("Index", "Login");
+            }
+          
+            if (viewBranches.ImageUpload == null || viewBranches.ImageUpload.ContentLength == 0)
+            {
+                ModelState.AddModelError("ImageUpload", "This field is required");
+            }
+            else if (!AppConstants.validImageTypes.Contains(viewBranches.ImageUpload.ContentType))
+            {
+                ModelState.AddModelError("ImageUpload", "Please choose either a GIF, JPG or PNG image.");
             }
 
-            ViewBag.RestaurantId = new SelectList(db.Restaurants, "Id", "Name", branchinfo.RestaurantId);
-            return View(branchinfo);
+            if (ModelState.IsValid)
+            {
+                var branchinfo = new BranchInfo
+                {
+                    Name = viewBranches.Name,
+                    Address = viewBranches.Address,
+                    RestaurantId = viewBranches.RestaurantId
+                };
+
+                if (viewBranches.ImageUpload != null && viewBranches.ImageUpload.ContentLength > 0)
+                {
+                    const string uploadDir = "~/Images";
+                    var imagePath = Path.Combine(Server.MapPath(uploadDir), viewBranches.ImageUpload.FileName);
+                    var imageUrl = Path.Combine(uploadDir, viewBranches.ImageUpload.FileName);
+                    if (!System.IO.File.Exists(imagePath))
+                    {
+                        viewBranches.ImageUpload.SaveAs(imagePath);
+                        branchinfo.ImageUrl = imageUrl;
+                        db.Branches.Add(branchinfo);
+                        await db.SaveChangesAsync();
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("ImageUpload", "Image already Exist");
+                    }
+                }
+            }
+
+            ViewBag.RestaurantId = new SelectList(db.Restaurants, "Id", "Name", viewBranches.RestaurantId);
+            return View(viewBranches);
         }
 
         // GET: /Branches/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -84,8 +139,12 @@ namespace RestuarantApp_API.AdminController
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include="Id,Name,Address,ImageUrl,RestaurantId")] BranchInfo branchinfo)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Name,Address,ImageUrl,RestaurantId")] BranchInfo branchinfo)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(branchinfo).State = EntityState.Modified;
@@ -99,6 +158,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Branches/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -116,6 +179,10 @@ namespace RestuarantApp_API.AdminController
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             BranchInfo branchinfo = await db.Branches.FindAsync(id);
             db.Branches.Remove(branchinfo);
             await db.SaveChangesAsync();
