+++ Restaurant_API/RestuarantApp_API/AdminController/CategoriesController.cs
@@ -9,6 +9,10 @@ using System.Web;
 using System.Web.Mvc;
 using RestuarantApp_API.Models;
 using RestuarantApp_API.Context;
+using RestuarantApp_API.Dtos;
+using Newtonsoft.Json;
+using RestuarantApp_API.Constants;
+using System.IO;
 
 namespace RestuarantApp_API.AdminController
 {
@@ -19,12 +23,20 @@ namespace RestuarantApp_API.AdminController
         // GET: /Categories/
         public async Task<ActionResult> Index()
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             return View(await db.Categories.ToListAsync());
         }
 
         // GET: /Categories/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -40,6 +52,11 @@ namespace RestuarantApp_API.AdminController
         // GET: /Categories/Create
         public ActionResult Create()
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            ViewBag.SubCategories = db.SubCategories.ToList();
             return View();
         }
 
@@ -50,6 +67,10 @@ namespace RestuarantApp_API.AdminController
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include="Id,Name,ImageUrl,CreatedDate,CreatedBy")] Category category)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (ModelState.IsValid)
             {
                 db.Categories.Add(category);
@@ -59,10 +80,112 @@ namespace RestuarantApp_API.AdminController
 
             return View(category);
         }
+        //[AcceptVerbs(HttpVerbs.Post)]
+        //public async Task<JsonResult> CreateCategories(Category category)
+        //{
+        //    if (Session["UserName"] == null)
+        //    {
+        //        return Json("session out");
+        //    }
+        //    var records = new List<SubCategory>();
+        //    if (Session["UserName"] == null)
+        //    {
+        //        return Json("session out");
+        //    }
+        //    if (category.SubCategoryIds != string.Empty && category.SubCategoryIds != null)
+        //    {
+        //        IEnumerable<int> ids = category.SubCategoryIds.Split(',').Select(str => int.Parse(str));
+        //        records = db.SubCategories.Where(rec => ids.Contains(rec.Id)).ToList();
+        //    }
 
+        //    foreach (var item in records)
+        //    {
+        //        category.SubCategories.Add(item);
+        //    }
+
+        //    string respponse = string.Empty;
+        //    if (ModelState.IsValid)
+        //    {
+        //        db.Categories.Add(category);
+        //        await db.SaveChangesAsync();
+        //    }
+        //    else
+        //    {
+        //        return Json(false);
+        //    }
+
+        //    ViewBag.SubCategoryId = new SelectList(db.SubCategories, "Id", "Name", category.SubCategoryId);
+        //    return Json(true);
+        //}
+
+        [AcceptVerbs(HttpVerbs.Post)]
+        public async Task<JsonResult> CreateCategories(HttpPostedFileBase[] files)
+        {
+            if (Session["UserName"] == null)
+            {
+                return Json("session out");
+            }
+            Category category = JsonConvert.DeserializeObject<Category>(Request["categories"]);
+            foreach (var file in files)
+            {
+                if (file == null || file.ContentLength == 0)
+                {
+                    //ModelState.AddModelError("ImageUpload", "This field is required");
+                    return Json("ImageUrl is required");
+                }
+                else if (!AppConstants.validImageTypes.Contains(file.ContentType))
+                {
+                    // ModelState.AddModelError("ImageUpload", "Please choose either a GIF, JPG or PNG image.");
+                    return Json("Please choose either a GIF, JPG or PNG image.");
+                }
+
+                if (file != null && file.ContentLength > 0)
+                {
+                    const string uploadDir = "~/Images";
+                    var imagePath = Path.Combine(Server.MapPath(uploadDir), file.FileName);
+                    var imageUrl = Path.Combine(uploadDir, file.FileName);
+                    if (!System.IO.File.Exists(imagePath))
+                    {
+                        file.SaveAs(imagePath);
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("ImageUpload", "Image already Exist");
+                        return Json("Image already Exist");
+                    }
+                }
+            }
+            var records = new List<SubCategory>();
+            if (category.SubCategoryIds != string.Empty && category.SubCategoryIds != null)
+            {
+                IEnumerable<int> ids = category.SubCategoryIds.Split(',').Select(str => int.Parse(str));
+                records = db.SubCategories.Where(rec => ids.Contains(rec.Id)).ToList();
+            }
+            foreach (var item in records)
+            {
+                category.SubCategories.Add(item);
+            }
+            string respponse = string.Empty;
+            if (ModelState.IsValid)
+            {
+                db.Categories.Add(category);
+                await db.SaveChangesAsync();
+            }
+            else
+            {
+                return Json(false);
+            }
+
+           // ViewBag.SubCategoryId = new SelectList(db.SubCategories, "Id", "Name", menu.SubCategoryId);
+            return Json(true);
+        }
         // GET: /Categories/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -72,6 +195,13 @@ namespace RestuarantApp_API.AdminController
             {
                 return HttpNotFound();
             }
+            ViewBag.SubCategoryIds = category.SubCategories.Select(t => new SubCategoryDto
+            {
+                Id = t.Id,
+                Name = t.Name,
+                ImageUrl = t.ImageUrl
+            }).ToList();
+            ViewBag.SubCategories = db.SubCategories.ToList();
             return View(category);
         }
 
@@ -82,9 +212,36 @@ namespace RestuarantApp_API.AdminController
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include="Id,Name,ImageUrl,CreatedDate,CreatedBy")] Category category)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            var subCategories = new List<SubCategory>();
             if (ModelState.IsValid)
             {
                 db.Entry(category).State = EntityState.Modified;
+
+                var subCatIds = Request["SubCategoryIds"].ToString().Split(',');
+                if (subCatIds.Count() > 0)
+                {
+                    IEnumerable<int> ids = subCatIds.Select(str => int.Parse(str));
+                    subCategories = db.SubCategories.Where(rec => ids.Contains(rec.Id)).ToList();
+                }
+                db.Entry(category).Collection("SubCategories").Load();
+                foreach (var item in category.SubCategories.ToList())
+                {
+                    category.SubCategories.Remove(item);
+
+                }
+                foreach (var item in subCategories)
+                {
+                    category.SubCategories.Add(item);
+                }
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
@@ -94,6 +251,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Categories/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -111,6 +272,10 @@ namespace RestuarantApp_API.AdminController
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             Category category = await db.Categories.FindAsync(id);
             db.Categories.Remove(category);
             await db.SaveChangesAsync();
