+++ Restaurant_API/RestuarantApp_API/AdminController/MenusController.cs
@@ -9,6 +9,10 @@ using System.Web;
 using System.Web.Mvc;
 using RestuarantApp_API.Models;
 using RestuarantApp_API.Context;
+using Newtonsoft.Json;
+using RestuarantApp_API.Dtos;
+using RestuarantApp_API.Constants;
+using System.IO;
 
 namespace RestuarantApp_API.AdminController
 {
@@ -19,6 +23,11 @@ namespace RestuarantApp_API.AdminController
         // GET: /Menus/
         public async Task<ActionResult> Index()
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             var menus = db.Menus.Include(m => m.SubCategory);
             return View(await menus.ToListAsync());
         }
@@ -26,6 +35,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Menus/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -41,6 +54,12 @@ namespace RestuarantApp_API.AdminController
         // GET: /Menus/Create
         public ActionResult Create()
         {
+
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            ViewBag.MenuItems = db.MenuItems.ToList();
             ViewBag.SubCategoryId = new SelectList(db.SubCategories, "Id", "Name");
             return View();
         }
@@ -50,8 +69,12 @@ namespace RestuarantApp_API.AdminController
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include="Id,Name,Price,ImageUrl,Description,CreatedDate,CreatedBy,SubCategoryId")] Menu menu)
+        public async Task<ActionResult> Create([Bind(Include = "Id,Name,Price,ImageUrl,Description,CreatedDate,CreatedBy,SubCategoryId")] Menu menu)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (ModelState.IsValid)
             {
                 db.Menus.Add(menu);
@@ -66,6 +89,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Menus/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -75,7 +102,16 @@ namespace RestuarantApp_API.AdminController
             {
                 return HttpNotFound();
             }
+
             ViewBag.SubCategoryId = new SelectList(db.SubCategories, "Id", "Name", menu.SubCategoryId);
+            ViewBag.MenuItemIds = menu.MenuItems.Select(t => new MenuItemDto
+            {
+                Id = t.Id,
+                Name = t.Name,
+                Description = t.Description,
+                ImageUrl = t.ImageUrl
+            }).ToList();
+            ViewBag.MenuItems = db.MenuItems.ToList();
             return View(menu);
         }
 
@@ -84,25 +120,92 @@ namespace RestuarantApp_API.AdminController
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit([Bind(Include="Id,Name,Price,ImageUrl,Description,CreatedDate,CreatedBy,SubCategoryId")] Menu menu)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Name,Price,ImageUrl,Description,CreatedDate,CreatedBy,SubCategoryId,MenuItemsIds")] Menu menu)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            var menuItems = new List<MenuItem>();
             if (ModelState.IsValid)
             {
                 db.Entry(menu).State = EntityState.Modified;
+
+                var ItemsIds = Request["MenuItemsIds"].ToString().Split(',');
+                if (ItemsIds.Count() > 0)
+                {
+                    IEnumerable<int> ids = ItemsIds.Select(str => int.Parse(str));
+                    menuItems = db.MenuItems.Where(rec => ids.Contains(rec.Id)).ToList();
+                }
+                db.Entry(menu).Collection("MenuItems").Load();
+                foreach (var item in menu.MenuItems.ToList())
+                {
+                    menu.MenuItems.Remove(item);
+                   
+                }
+                foreach (var item in menuItems)
+                {
+                    menu.MenuItems.Add(item);
+                }
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
             ViewBag.SubCategoryId = new SelectList(db.SubCategories, "Id", "Name", menu.SubCategoryId);
             return View(menu);
         }
+
         [AcceptVerbs(HttpVerbs.Post)]
-        public async Task<JsonResult> CreateMenu(Menu menu)
+        public async Task<JsonResult> CreateMenu(HttpPostedFileBase[] files)
         {
+            if (Session["UserName"] == null)
+            {
+                return Json("session out");
+            }
+            Menu menu = JsonConvert.DeserializeObject<Menu>(Request["menu"]);
+            foreach (var file in files)
+            {
+                if (file == null || file.ContentLength == 0)
+                {
+                    //ModelState.AddModelError("ImageUpload", "This field is required");
+                    return Json("ImageUrl is required");
+                }
+                else if (!AppConstants.validImageTypes.Contains(file.ContentType))
+                {
+                   // ModelState.AddModelError("ImageUpload", "Please choose either a GIF, JPG or PNG image.");
+                    return Json("Please choose either a GIF, JPG or PNG image.");
+                }
+
+                if (file != null && file.ContentLength > 0)
+                {
+                    const string uploadDir = "~/Images";
+                    var imagePath = Path.Combine(Server.MapPath(uploadDir), file.FileName);
+                    var imageUrl = Path.Combine(uploadDir, file.FileName);
+                    if (!System.IO.File.Exists(imagePath))
+                    {
+                        file.SaveAs(imagePath);
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("ImageUpload", "Image already Exist");
+                        return Json("Image already Exist");
+                    }
+                }
+            }
+            var records = new List<MenuItem>();
+            if (menu.MenuItemsIds != string.Empty && menu.MenuItemsIds != null)
+            {
+                IEnumerable<int> ids = menu.MenuItemsIds.Split(',').Select(str => int.Parse(str));
+                records = db.MenuItems.Where(rec => ids.Contains(rec.Id)).ToList();
+            }
+            foreach (var item in records)
+            {
+                menu.MenuItems.Add(item);
+            }
             string respponse = string.Empty;
             if (ModelState.IsValid)
             {
-                 db.Menus.Add(menu);
-                 await db.SaveChangesAsync();
+                db.Menus.Add(menu);
+                await db.SaveChangesAsync();
             }
             else
             {
@@ -115,6 +218,10 @@ namespace RestuarantApp_API.AdminController
         // GET: /Menus/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -132,12 +239,30 @@ namespace RestuarantApp_API.AdminController
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (Session["UserName"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             Menu menu = await db.Menus.FindAsync(id);
             db.Menus.Remove(menu);
             await db.SaveChangesAsync();
             return RedirectToAction("Index");
         }
+        public void FileUpload(HttpPostedFileBase file)
+        {
 
+            if (file != null)
+            {
+                    string ImageName = System.IO.Path.GetFileName(file.FileName);
+                    string physicalPath = Server.MapPath("~/images/" + ImageName);
+
+                    if (!System.IO.File.Exists(physicalPath))
+                {
+                    // save image in folder
+                    file.SaveAs(physicalPath);
+                }
+            }
+        }
         protected override void Dispose(bool disposing)
         {
             if (disposing)
@@ -146,5 +271,8 @@ namespace RestuarantApp_API.AdminController
             }
             base.Dispose(disposing);
         }
+
+
     }
+  
 }
